+++ src/macros/mod.rs
@@ -2,3 +2,4 @@
 
 pub mod matlab_macro;
 pub mod r_macro;
+pub mod julia_macro;